♻️ mock data 리팩토링
